warning cleanup
Cleaned up warning reported by the dashboard. refactored warning
supression pragma to be used as locally as possible. Added FTLE
filter, moved the code from the sq vortex detect filter.

+++ Plugins/SciberQuestToolKit/WindowsSystemInterface.cxx
@@ -92,7 +92,7 @@ unsigned long WindowsSystemInterface::GetMemoryUsed()
     cerr << "Failed to obtain memory information." << endl;
     return 0;
     }
-  return pmc.WorkingSetSize/1024;
+  return (unsigned long)(pmc.WorkingSetSize/1024);
 }
 
 //-----------------------------------------------------------------------------
